# docs/swagger.yaml
basePath: /
definitions:
  web.IndikatorPokinOpdCreateRequest:
    properties:
      indikator:
        type: string
      tujuan_pokin_opd_id:
        type: integer
    type: object
  web.IndikatorPokinOpdResponse:
    properties:
      id:
        type: integer
      indikator:
        type: string
    type: object
  web.IndikatorPokinOpdStrategicCreateRequest:
    properties:
      indikator:
        type: string
      pokin_opd_strategic_id:
        type: integer
    required:
    - pokin_opd_strategic_id
    type: object
  web.IndikatorPokinOpdStrategicResponse:
    properties:
      id:
        type: integer
      indikator:
        type: string
      pokin_opd_strategic_id:
        type: integer
    type: object
  web.IndikatorPokinOpdStrategicUpdateRequest:
    properties:
      id:
        type: integer
      indikator:
        type: string
      pokin_opd_strategic_id:
        type: integer
    required:
    - id
    - pokin_opd_strategic_id
    type: object
...
  web.IndikatorPokinOpdUpdateRequest:
    properties:
      id:
        type: integer
      indikator:
        type: string
    required:
    - id
    type: object
...
  web.WebResponse:
    description: Web Response
    properties:
      code:
        type: integer
      data: {}
      status:
        type: string
    type: object
host: ${PROD_HOSTNAME}
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API For Alur Kerja Services
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Alur Kerja Service API
  version: "1.0"
paths:
  /indikator-pokin-opd-strategics:
    get:
      consumes:
      - application/json
      description: Get list of all Indikator Pokin OPD Strategic
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.IndikatorPokinOpdStrategicResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: List All Indikator Pokin OPD Strategic
      tags:
      - Indikator Pokin OPD Strategic
    post:
      consumes:
      - application/json
      description: Create new Indikator Pokin OPD Strategic
      parameters:
      - description: Indikator Pokin OPD Strategic Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.IndikatorPokinOpdStrategicCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdStrategicResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Indikator Pokin OPD Strategic
      tags:
      - Indikator Pokin OPD Strategic
  /indikator-pokin-opd-strategics/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing Indikator Pokin OPD Strategic by ID
      parameters:
      - description: Indikator Pokin OPD Strategic ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.WebResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Indikator Pokin OPD Strategic
      tags:
      - Indikator Pokin OPD Strategic
    get:
      consumes:
      - application/json
      description: Get Indikator Pokin OPD Strategic detail by ID
      parameters:
      - description: Indikator Pokin OPD Strategic ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdStrategicResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Indikator Pokin OPD Strategic by ID
      tags:
      - Indikator Pokin OPD Strategic
    put:
      consumes:
      - application/json
      description: Update existing Indikator Pokin OPD Strategic by ID
      parameters:
      - description: Indikator Pokin OPD Strategic ID
        in: path
        name: id
        required: true
        type: integer
      - description: Indikator Pokin OPD Strategic Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.IndikatorPokinOpdStrategicUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdStrategicResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Indikator Pokin OPD Strategic
      tags:
      - Indikator Pokin OPD Strategic
...
  /indikator-pokin-opds:
    get:
      consumes:
      - application/json
      description: Get list of all Indikator Pokin Opd
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/web.IndikatorPokinOpdResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: List All Indikator Pokin Opd
      tags:
      - Indikator Pokin Opd
    post:
      consumes:
      - application/json
      description: Create new Indikator Pokin Opd
      parameters:
      - description: Indikator Pokin Opd Create Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.IndikatorPokinOpdCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Create Indikator Pokin Opd
      tags:
      - Indikator Pokin Opd
  /indikator-pokin-opds/{id}:
    delete:
      consumes:
      - application/json
      description: Delete existing Indikator Pokin Opd by ID
      parameters:
      - description: Indikator Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Delete Indikator Pokin Opd
      tags:
      - Indikator Pokin Opd
    get:
      consumes:
      - application/json
      description: Get Indikator Pokin Opd detail by ID
      parameters:
      - description: Indikator Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Get Indikator Pokin Opd by ID
      tags:
      - Indikator Pokin Opd
    put:
      consumes:
      - application/json
      description: Update existing Indikator Pokin Opd by ID
      parameters:
      - description: Indikator Pokin Opd ID
        in: path
        name: id
        required: true
        type: integer
      - description: Indikator Pokin Opd Update Request
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/web.IndikatorPokinOpdUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.WebResponse'
            - properties:
                data:
                  $ref: '#/definitions/web.IndikatorPokinOpdResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/web.WebResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/web.WebResponse'
      summary: Update Indikator Pokin Opd
      tags:
      - Indikator Pokin Opd
...
schemes:
- http
- https
swagger: "2.0"